tests: speedtest: deactivated & fixed pylint errors
