+++ tests_deact/speedtest/test_mf_class.py
@@ -8,7 +8,7 @@ mf = miniflask.init(
 )
 
 
-class event():
+class event():  # pylint: disable=too-few-public-methods
 
     @classmethod
     def func(cls, x):
@@ -20,5 +20,5 @@ def test_mf_class():
     mf.load("moduleunique")
 
     a = 0
-    for i in range(10000000):
+    for _ in range(10000000):
         a += event.func(42)
