+++ tests_deact/speedtest/test_mf_event_nomfargs.py
@@ -13,5 +13,5 @@ def test_mf_event_nomfargs():
     mf.load("modulenomfargs")
 
     a = 0
-    for i in range(10000000):
+    for _ in range(10000000):
         a += event.func(42)
